Update README.md
This repository contains the final project for my CodeFirstGirls Coding Kickstarter - Introduction to Data Science & SQL course. It features a relational database designed to manage and analyse information for a STEM tuition institution, such as tutor and student information, course registrations, payments, and more. The project also demonstrates the use of SQL queries, views, stored functions, and procedures to perform various tasks and analyses on the data.